Move compiler detection header
