+++ include/lucenaBAL/details/lbalDetectCompiler.hpp
